[feat/#171] 길찾기 Util 코드 구현
