+++ Solply/Solply/Presentation/Place/PlaceDetail/Effect/PlaceDetailEffect.swift
@@ -11,30 +11,20 @@ struct PlaceDetailEffect {
     private let courseService = CourseService()
     private let placeService = PlaceService()
     
-    func openNaverMapRoute(
+    func findDirection(
         startLatitude: Double,
         startLongitude: Double,
         destinationLatitude: Double,
         destinationLongitude: Double,
         destinationName: String
     ) {
-        let schemeURLString = "nmap://route/public?" +
-        "sname=내 위치&slat=\(startLatitude)&slng=\(startLongitude)" +
-        "&dname=\(destinationName)&dlat=\(destinationLatitude)&dlng=\(destinationLongitude)" +
-        "&appname=\(Bundle.main.bundleIdentifier ?? "")"
-        
-        guard let encodedSchemeURLString = schemeURLString.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed),
-              let schemeURL = URL(string: encodedSchemeURLString) else {
-            return
-        }
-        
-        if UIApplication.shared.canOpenURL(schemeURL) {
-            UIApplication.shared.open(schemeURL)
-        } else {
-            if let appStoreURL = URL(string: "https://apps.apple.com/app/id311867728") {
-                UIApplication.shared.open(appStoreURL)
-            }
-        }
+        NaverMapRouteHelper.findDirection(
+            startLatitude: startLatitude,
+            startLongitude: startLongitude,
+            destinationLatitude: destinationLatitude,
+            destinationLongitude: destinationLongitude,
+            destinationName: destinationName
+        )
     }
 
     func fetchCourseArchive(townId: Int, placeId: Int?) async -> PlaceDetailAction {
